wallet: split wallet store unit tests to specific files
